feat(T09): adding confirm modal

+++ src/components/TodoLayout/TodoList/components/Todo/Todo.tsx
@@ -1,7 +1,10 @@
+import { NotificationType } from "@/common/constants/NotificationConstants.ts";
+import { TodoConfirmMessages, TodoNotificationMessages } from "@/common/constants/TodoConstants.ts";
 import { Todo as TodoProps } from "@/common/types/Todo.ts";
 import { Input } from "@/components/ui/Input/Input.tsx";
 import { useAppDispatch } from "@/hooks/useAppDispatch.ts";
 import { useModalState } from "@/hooks/useModalState.ts";
+import { setNotification } from "@/store/actions/notificationActionCreators.ts";
 import { deleteTodo, setCurrentTodo, updateStatusTodo } from "@/store/actions/todoActionCreators.ts";
 import cn from "classnames";
 import { FC, useState } from "react";
@@ -13,7 +16,8 @@ import styles from "./Todo.module.scss";
 
 export const Todo: FC<TodoProps> = ({ todoTitle, createdDate, expirationDate, todoId, isCompleted }) => {
   const [isShowInfo, setIsShowInfo] = useState(false);
-  const setModalActive = useModalState("editTodoModal")[1];
+  const setEditModalActive = useModalState("editTodoModal")[1];
+  const setConfirmModalActive = useModalState("modalConfirm")[1];
   const dispatch = useAppDispatch();
 
   const handleChangeStatusTodo = () => {
@@ -21,12 +25,18 @@ export const Todo: FC<TodoProps> = ({ todoTitle, createdDate, expirationDate, to
   };
 
   const handleClickDeleteTodo = () => {
-    dispatch(deleteTodo(todoId));
+    setConfirmModalActive(true, {
+      confirmCallback: () => {
+        dispatch(deleteTodo(todoId));
+        dispatch(setNotification({ title: TodoNotificationMessages.DELETE_TODO, type: NotificationType.SUCCESS }));
+      },
+      message: TodoConfirmMessages.DELETE_TODO,
+    });
   };
 
   const handleClickEditTodo = () => {
     if (!isCompleted) {
-      setModalActive(true);
+      setEditModalActive(true);
       dispatch(setCurrentTodo({ todoId, todoTitle, expirationDate, createdDate, isCompleted }));
     }
   };
